colour_math::fdrn: implement * and / by literal integers (vs u8)

+++ colour_math/src/beigui/hue_wheel.rs
@@ -244,7 +244,7 @@ pub trait Graticule {
         draw_shapes.set_line_colour(&HCV::WHITE); // * UFDRNumber::from(0.25));
         let centre = Point::default();
         for num in 1..=num_rings {
-            let radius: UFDRNumber = UFDRNumber::ONE * num / num_rings;
+            let radius: UFDRNumber = UFDRNumber::ONE * num as i32 / num_rings as i32;
             draw_shapes.draw_circle(centre, radius * zoom.scale(), false);
         }
     }
